+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @to-skills/cli
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @to-skills/core@0.15.0
+
 ## 0.3.0
 
 ### Minor Changes